finishing

+++ migrations/postgres/009_weather_history.sql
@@ -3,7 +3,7 @@ CREATE TABLE weather_history(
     id BIGSERIAL PRIMARY KEY,
     user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     city VARCHAR(255) NOT NULL REFERENCES cities(city) ON DELETE CASCADE,
-    temperature INT NOT NULL,
+    temperature DOUBLE NOT NULL,
     description VARCHAR(255) NOT NULL,
     requested_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
